split GCR lookup from docker package
