+++ v2/container/gcr/project.go
@@ -1,4 +1,4 @@
-package uri
+package gcr
 
 import (
 	"fmt"
